requirements.txt
ajax==3.0.0
asgiref==3.4.1
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.4
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
python-decouple==3.5
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
